pyproject.toml
[project]
name = "stock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.2.8",
    "django-cors-headers>=4.9.0",
    "django-tenants==3.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2-binary>=2.9.11",
]
